feat: [SCRUM-154] prevent unnecessary rerender
